+++ backend/api/season.py
@@ -1,7 +1,6 @@
 from flask import Blueprint, jsonify
-from services.store.storage import DbCursor
 from services.store.season import get_season, list_season_ids, insert_season, update_season, delete_season
-from services.store.field import delete_field_ndvi_raster
+from services.store.ndvi_raster import get_ndvi_raster
 from api.authentication import authentication_required
 from config import NDVI
 import os
@@ -12,7 +11,7 @@ api = Blueprint("season", __name__, url_prefix="/season")
 
 @api.route("/<int:field_id>", methods=["GET"])
 @authentication_required
-def list_all_season_options(user_id, __, field_id):
+def retrieve_season_options(user_id, _, field_id):
     season_ids = list_season_ids(user_id, field_id)
     if season_ids is not None:
         return season_ids, 200
@@ -53,18 +52,13 @@ def upgister_season(user_id, data, field_id, season_id):
 @api.route("/unregister/<int:field_id>/<season_id>", methods=["DELETE"])
 @authentication_required
 def unregister_season(user_id, _, field_id, season_id):
-    db_cursor = DbCursor()
-    deleted_ndvi_raster = None
-    with db_cursor as cursor:
-        delete_season(user_id, field_id, season_id, cursor=cursor)
-        deleted_ndvi_raster = delete_field_ndvi_raster(field_id, season_id)
-    if db_cursor.error is None:
-        if deleted_ndvi_raster is not None:
-            try:
-                os.remove(os.path.join(NDVI.data_folder, deleted_ndvi_raster))
-            except Exception as error:
-                print("[Season API]", error)
-                return jsonify({"data": "Failed to unregister the associated NDVI raster"}), 500
+    deleted_ndvi_raster = get_ndvi_raster(user_id, field_id, season_id)
+    if deleted_ndvi_raster is not None and delete_season(user_id, field_id, season_id):
+        try:
+            os.remove(os.path.join(NDVI.data_folder, deleted_ndvi_raster))
+        except Exception as error:
+            print("[Season API]", error)
+            return jsonify({"data": "Failed to unregister the associated NDVI raster"}), 500
         return jsonify({"data": "Successfully unregister the season"}), 204
     else:
         return jsonify({"data": "Failed to unregister the season"}), 500
